requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astroid==3.3.5
autopep8==2.3.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
Django==5.0.9
django-allauth==65.2.0
django-mssql-backend==2.8.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
fastapi==0.115.2
flake8==7.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mssql-django==1.5
mypy==1.12.0
mypy-extensions==1.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pylint==3.3.1
pyodbc==5.1.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.40.0
tomlkit==0.13.2
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
whitenoise==6.7.0
